refactor: finish MidStreamInjectionHookInstaller extraction (Step 29)
Move the 5 remaining hook-installation methods out of orchestrator.py into the
MidStreamInjectionHookInstaller collaborator, completing the roadmap's deferred
Step 29. A1 (prior commit) was the callback-unification prerequisite; with the
two injection closures already unified, these now extract cleanly:

- _setup_hook_manager_for_agent (dispatcher)
- _setup_native_hooks_for_agent
- _setup_codex_mcp_hooks
- _setup_codex_hybrid_hooks
- _register_round_timeout_hooks

Each becomes a thin orchestrator delegator. Cross-method calls between the moved
methods route through orch._<delegator> to preserve test monkeypatch-safety.
_codex_mcp_hook_agents is still written on the orchestrator (a test reads it there).

orchestrator.py 8,422 -> 7,910 lines. Removed the now-unused hook imports.

Validated: hooks/restart-and-external-tools/broadcast-subagents integration suites,
the 37-test characterization safety net, and the injection/decomposition/novelty
suites all green. Only pre-existing test_subagent_round_timeouts FileNotFound
failures remain (unrelated, confirmed identical on baseline).
