+++ massgen/orchestrator.py
@@ -59,14 +59,7 @@ from .logger_config import (
 from .mcp_tools.hooks import (
     BackgroundToolCompleteHook,
     GeneralHookManager,
-    HighPriorityTaskReminderHook,
-    HookType,
     HumanInputHook,
-    MediaCallLedgerHook,
-    MidStreamInjectionHook,
-    PythonCallableHook,
-    RoundTimeoutPostHook,
-    RoundTimeoutPreHook,
     RoundTimeoutState,
     SubagentCompleteHook,
 )
@@ -3482,171 +3475,8 @@ class Orchestrator(ChatAgent):
         agent: ChatAgent,
         answers: dict[str, str],
     ) -> None:
-        """Set up hooks for agent - uses native adapter for Claude Code, GeneralHookManager for others.
-
-        This routes hook setup based on backend capabilities:
-        - Backends with native hook support (Claude Code): Use NativeHookAdapter
-        - Standard backends: Use GeneralHookManager
-
-        Both paths set up the same hooks:
-        1. MidStreamInjectionHook - injects answers from other agents into tool results
-        2. HighPriorityTaskReminderHook - reminds to document high-priority task completions (round mode only)
-
-        Args:
-            agent_id: The agent identifier
-            agent: The ChatAgent instance
-            answers: Dict of existing answers when agent started (used to detect new answers)
-        """
-        # Runtime human input must work for all backends, including those that
-        # don't support hook registration (hookless fallback / Codex path).
-        self._ensure_runtime_human_input_hook_initialized()
-        self._ensure_runtime_inbox_poller_initialized()
-
-        backend = getattr(agent, "backend", None)
-        backend_provider = backend.get_provider_name() if backend and hasattr(backend, "get_provider_name") else ""
-
-        # Codex uses a hybrid path: native Bash hooks plus MCP/file-based payload delivery.
-        if (
-            backend_provider == "codex"
-            and hasattr(agent.backend, "supports_native_hooks")
-            and agent.backend.supports_native_hooks()
-            and hasattr(agent.backend, "supports_mcp_server_hooks")
-            and agent.backend.supports_mcp_server_hooks()
-        ):
-            self._setup_codex_hybrid_hooks(agent_id, agent, answers)
-            return
-
-        # Check if backend supports native hooks (e.g., Claude Code)
-        if hasattr(agent.backend, "supports_native_hooks") and agent.backend.supports_native_hooks():
-            self._setup_native_hooks_for_agent(agent_id, agent, answers)
-            return
-
-        # Check if backend supports MCP server-level hooks (e.g., Codex)
-        if hasattr(agent.backend, "supports_mcp_server_hooks") and agent.backend.supports_mcp_server_hooks():
-            self._setup_codex_mcp_hooks(agent_id, agent, answers)
-            return
-
-        # Fall back to GeneralHookManager for standard backends
-        if not hasattr(agent.backend, "set_general_hook_manager"):
-            return
-
-        # Create hook manager
-        manager = GeneralHookManager()
-
-        # Create mid-stream injection hook with closure-based callback
-        mid_stream_hook = MidStreamInjectionHook()
-
-        # Define the injection callback (captures agent_id and answers).
-        # A1: both the GeneralHookManager and native paths route through the
-        # single unified installer method so they can no longer drift.
-        async def get_injection_content() -> str | None:
-            return await self._build_midstream_injection(agent_id, answers, native=False)
-
-        # Set callback on hook
-        mid_stream_hook.set_callback(get_injection_content)
-
-        # Register mid-stream injection hook first (maintains current behavior order)
-        manager.register_global_hook(HookType.POST_TOOL_USE, mid_stream_hook)
-
-        # Register high-priority task reminder hook (enabled round-wise when capture is enabled)
-        if self._is_round_learning_capture_enabled():
-            reminder_hook = HighPriorityTaskReminderHook()
-            manager.register_global_hook(HookType.POST_TOOL_USE, reminder_hook)
-
-        # Register media call ledger hook (read_media/generate_media provenance capture)
-        manager.register_global_hook(HookType.POST_TOOL_USE, MediaCallLedgerHook())
-
-        # Register human input hook (shared across all agents)
-        manager.register_global_hook(HookType.POST_TOOL_USE, self._human_input_hook)
-
-        # Register subagent completion hook for background result injection
-        if self._background_subagents_enabled:
-            subagent_hook = SubagentCompleteHook(
-                injection_strategy=self._background_subagent_injection_strategy,
-            )
-
-            # Create a closure that captures agent_id for pending results retrieval
-            def make_pending_getter(aid: str):
-                return lambda: self._get_pending_subagent_results_async(aid)
-
-            subagent_hook.set_pending_results_getter(make_pending_getter(agent_id))
-            manager.register_global_hook(HookType.POST_TOOL_USE, subagent_hook)
-            logger.debug(f"[Orchestrator] Registered SubagentCompleteHook for {agent_id}")
-
-            # Wire background tool delegate so list/status/result/cancel route to subagents
-            if hasattr(agent.backend, "register_background_delegate"):
-                from massgen.subagent.background_delegate import (
-                    SubagentBackgroundDelegate,
-                )
-
-                def _make_call_tool(aid: str):
-                    return lambda tool_name, params: self._call_subagent_mcp_tool_async(
-                        aid,
-                        tool_name,
-                        params,
-                    )
-
-                delegate = SubagentBackgroundDelegate(
-                    call_tool=_make_call_tool(agent_id),
-                    agent_id=agent_id,
-                )
-                agent.backend.register_background_delegate(delegate)
-                logger.debug(f"[Orchestrator] Registered SubagentBackgroundDelegate for {agent_id}")
-
-        # Register background tool completion hook for async tool result injection
-        if hasattr(agent.backend, "get_pending_background_tool_results"):
-            background_tool_hook = BackgroundToolCompleteHook()
-            background_tool_hook.set_completed_jobs_getter(
-                agent.backend.get_pending_background_tool_results,
-            )
-            manager.register_global_hook(HookType.POST_TOOL_USE, background_tool_hook)
-            logger.debug(
-                f"[Orchestrator] Registered BackgroundToolCompleteHook for {agent_id}",
-            )
-        # Register per-round timeout hooks if configured
-        self._register_round_timeout_hooks(agent_id, manager)
-
-        # Register user-configured hooks from agent backend config
-        if hasattr(agent.backend, "config") and agent.backend.config:
-            agent_hooks = agent.backend.config.get("hooks")
-            if agent_hooks:
-                manager.register_hooks_from_config(agent_hooks, agent_id=agent_id)
-                logger.debug(
-                    f"[Orchestrator] Registered user-configured hooks for {agent_id}",
-                )
-
-        # Set manager on backend
-        agent.backend.set_general_hook_manager(manager)
-        if hasattr(agent.backend, "set_background_wait_interrupt_provider"):
-
-            async def _wait_interrupt_provider(
-                requested_agent_id: str,
-                *,
-                _agent_id: str = agent_id,
-            ) -> dict[str, Any] | None:
-                target_agent_id = requested_agent_id or _agent_id
-                if hasattr(self, "cancellation_manager") and self.cancellation_manager and self.cancellation_manager.is_cancelled:
-                    return {
-                        "interrupt_reason": "turn_cancelled",
-                        "injected_content": None,
-                    }
-
-                runtime_sections = await self._collect_no_hook_runtime_fallback_sections(
-                    target_agent_id,
-                )
-                if not runtime_sections:
-                    return None
-                return {
-                    "interrupt_reason": "runtime_injection_available",
-                    "injected_content": "\n".join(runtime_sections),
-                }
-
-            agent.backend.set_background_wait_interrupt_provider(
-                _wait_interrupt_provider,
-            )
-        logger.debug(
-            f"[Orchestrator] Set up hook manager for {agent_id} with mid-stream and reminder hooks",
-        )
+        """Route hook setup by backend capability (delegates to MidStreamInjectionHookInstaller)."""
+        self._midstream_injection_hook_installer.setup_hook_manager_for_agent(agent_id, agent, answers)
 
     def _setup_codex_mcp_hooks(
         self,
@@ -3654,54 +3484,8 @@ class Orchestrator(ChatAgent):
         agent: ChatAgent,
         answers: dict[str, str],
     ) -> None:
-        """Set up MCP server-level hook delivery for Codex backends.
-
-        Instead of registering hooks on a GeneralHookManager, this stores a
-        reference so the streaming loop can call _flush_codex_hook_payloads()
-        to write injection files that the MCP middleware consumes.
-        """
-        # Mark this agent as using MCP server hooks
-        if not hasattr(self, "_codex_mcp_hook_agents"):
-            self._codex_mcp_hook_agents: dict[str, dict[str, Any]] = {}
-
-        self._codex_mcp_hook_agents[agent_id] = {
-            "agent": agent,
-            "answers": answers,
-        }
-
-        # Set up the background wait interrupt provider (reuse existing pattern)
-        if hasattr(agent.backend, "set_background_wait_interrupt_provider"):
-
-            async def _wait_interrupt_provider(
-                requested_agent_id: str,
-                *,
-                _agent_id: str = agent_id,
-            ) -> dict[str, Any] | None:
-                target_agent_id = requested_agent_id or _agent_id
-                if hasattr(self, "cancellation_manager") and self.cancellation_manager and self.cancellation_manager.is_cancelled:
-                    return {
-                        "interrupt_reason": "turn_cancelled",
-                        "injected_content": None,
-                    }
-
-                runtime_sections = await self._collect_no_hook_runtime_fallback_sections(
-                    target_agent_id,
-                )
-                if not runtime_sections:
-                    return None
-                return {
-                    "interrupt_reason": "runtime_injection_available",
-                    "injected_content": "\n".join(runtime_sections),
-                }
-
-            agent.backend.set_background_wait_interrupt_provider(
-                _wait_interrupt_provider,
-            )
-
-        logger.info(
-            "[Orchestrator] Set up MCP server-level hook delivery for %s",
-            agent_id,
-        )
+        """Set up Codex MCP server-level hook delivery (delegates to MidStreamInjectionHookInstaller)."""
+        self._midstream_injection_hook_installer.setup_codex_mcp_hooks(agent_id, agent, answers)
 
     def _setup_codex_hybrid_hooks(
         self,
@@ -3709,53 +3493,8 @@ class Orchestrator(ChatAgent):
         agent: ChatAgent,
         answers: dict[str, str],
     ) -> None:
-        """Set up Codex's hybrid delivery path.
-
-        Codex native hooks currently cover Bash-only ``PreToolUse`` and
-        ``PostToolUse``. MassGen runtime payload delivery still flows through the
-        shared ``hook_post_tool_use.json`` file and the MCP/file carry-forward
-        path, so we register a lightweight native Bash bridge and keep the
-        existing Codex MCP setup in place.
-        """
-        adapter = agent.backend.get_native_hook_adapter()
-        if not adapter:
-            logger.warning(
-                "[Orchestrator] Codex backend reported native hooks but no adapter was available for %s",
-                agent_id,
-            )
-            self._setup_codex_mcp_hooks(agent_id, agent, answers)
-            return
-
-        manager = GeneralHookManager()
-        manager.register_global_hook(
-            HookType.POST_TOOL_USE,
-            PythonCallableHook(
-                name="codex_post_tool_bridge",
-                handler=lambda _event: None,
-                matcher="Bash",
-            ),
-        )
-
-        native_config = adapter.build_native_hooks_config(
-            manager,
-            agent_id=agent_id,
-        )
-        agent.backend.set_native_hooks_config(native_config)
-        # Codex's native hook surface is Bash-only, but the TUI and manual
-        # wrap-up flow still need real timeout hook objects in agent state.
-        # Register those separately so request_answer_now() and timeout status
-        # work on the hybrid path without changing the native hooks config.
-        timeout_manager = GeneralHookManager()
-        self._register_round_timeout_hooks(agent_id, timeout_manager)
-        self._setup_codex_mcp_hooks(agent_id, agent, answers)
-
-        hooks = native_config.get("hooks", {}) if isinstance(native_config, dict) else {}
-        logger.info(
-            "[Orchestrator] Set up Codex hybrid hooks for %s: PreToolUse=%d, PostToolUse=%d",
-            agent_id,
-            len(hooks.get("PreToolUse", [])),
-            len(hooks.get("PostToolUse", [])),
-        )
+        """Set up Codex's hybrid delivery path (delegates to MidStreamInjectionHookInstaller)."""
+        self._midstream_injection_hook_installer.setup_codex_hybrid_hooks(agent_id, agent, answers)
 
     async def _collect_round_timeout_runtime_sections(
         self,
@@ -4470,97 +4209,8 @@ class Orchestrator(ChatAgent):
         agent_id: str,
         manager: GeneralHookManager,
     ) -> None:
-        """Register per-round timeout hooks if configured.
-
-        This creates two hooks:
-        1. RoundTimeoutPostHook (soft timeout) - Injects warning message after tool calls
-        2. RoundTimeoutPreHook (hard timeout) - Blocks non-terminal tools after grace period
-
-        The hooks are stored in agent_states so they can be reset when a new round starts.
-
-        Args:
-            agent_id: The agent identifier
-            manager: The GeneralHookManager to register hooks with
-        """
-        # Get timeout config
-        timeout_config = self.config.timeout_config
-        initial_timeout = timeout_config.initial_round_timeout_seconds
-        subsequent_timeout = timeout_config.subsequent_round_timeout_seconds
-        grace_seconds = timeout_config.round_timeout_grace_seconds
-
-        # Skip if no round timeouts configured
-        if initial_timeout is None and subsequent_timeout is None:
-            return
-
-        logger.info(
-            f"[Orchestrator] Registering round timeout hooks for {agent_id}: " f"initial={initial_timeout}s, subsequent={subsequent_timeout}s, grace={grace_seconds}s",
-        )
-
-        # Create closures that read from agent state
-        def get_round_start_time() -> float:
-            """Get the current round start time from agent state."""
-            start_time = self.agent_states[agent_id].round_start_time
-            if start_time is None:
-                # Fallback to current time if not set (shouldn't happen)
-                logger.warning(
-                    f"[Orchestrator] round_start_time is None for {agent_id}, using current time as fallback",
-                )
-                return time.time()
-            return start_time
-
-        def get_agent_round() -> int:
-            """Get the current round number from coordination tracker."""
-            return self.coordination_tracker.get_agent_round(agent_id)
-
-        # Create shared state for coordinating soft -> hard timeout progression
-        # This ensures hard timeout only fires AFTER soft timeout has been injected
-        timeout_state = RoundTimeoutState()
-
-        # Get two-tier workspace setting from coordination config
-        # Suppressed when write_mode is active (write_mode replaces the old two-tier structure)
-        coordination_config = getattr(self.config, "coordination_config", None)
-        write_mode = getattr(coordination_config, "write_mode", None) if coordination_config else None
-        use_two_tier_workspace = False
-        if not (write_mode and write_mode != "legacy"):
-            use_two_tier_workspace = bool(
-                getattr(coordination_config, "use_two_tier_workspace", False),
-            )
-
-        # Create soft timeout hook (POST_TOOL_USE - injects warning)
-        post_hook = RoundTimeoutPostHook(
-            name=f"round_timeout_soft_{agent_id}",
-            get_round_start_time=get_round_start_time,
-            get_agent_round=get_agent_round,
-            initial_timeout_seconds=initial_timeout,
-            subsequent_timeout_seconds=subsequent_timeout,
-            grace_seconds=grace_seconds,
-            agent_id=agent_id,
-            shared_state=timeout_state,
-            use_two_tier_workspace=use_two_tier_workspace,
-        )
-
-        # Create hard timeout hook (PRE_TOOL_USE - blocks non-terminal tools)
-        pre_hook = RoundTimeoutPreHook(
-            name=f"round_timeout_hard_{agent_id}",
-            get_round_start_time=get_round_start_time,
-            get_agent_round=get_agent_round,
-            initial_timeout_seconds=initial_timeout,
-            subsequent_timeout_seconds=subsequent_timeout,
-            grace_seconds=grace_seconds,
-            agent_id=agent_id,
-            shared_state=timeout_state,
-        )
-
-        # Register hooks
-        manager.register_global_hook(HookType.POST_TOOL_USE, post_hook)
-        manager.register_global_hook(HookType.PRE_TOOL_USE, pre_hook)
-
-        # Store hook references so we can reset them on new rounds
-        self.agent_states[agent_id].round_timeout_hooks = (post_hook, pre_hook)
-        # Store the shared state so we can check force_terminate in the orchestrator loop
-        self.agent_states[agent_id].round_timeout_state = timeout_state
-
-        logger.debug(f"[Orchestrator] Registered round timeout hooks for {agent_id}")
+        """Register per-round timeout hooks (delegates to MidStreamInjectionHookInstaller)."""
+        self._midstream_injection_hook_installer.register_round_timeout_hooks(agent_id, manager)
 
     def _setup_native_hooks_for_agent(
         self,
@@ -4568,180 +4218,8 @@ class Orchestrator(ChatAgent):
         agent: ChatAgent,
         answers: dict[str, str],
     ) -> None:
-        """Set up native hooks for backends that support them (e.g., Claude Code).
-
-        This converts MassGen hooks to the backend's native format using the
-        NativeHookAdapter interface. The hooks are then executed natively by
-        the backend rather than through MassGen's GeneralHookManager.
-
-        Args:
-            agent_id: The agent identifier
-            agent: The ChatAgent instance
-            answers: Dict of existing answers when agent started (used to detect new answers)
-        """
-        # Get the native hook adapter from the backend
-        adapter = agent.backend.get_native_hook_adapter()
-        if not adapter:
-            logger.warning(
-                f"[Orchestrator] Backend supports native hooks but adapter unavailable for {agent_id}",
-            )
-            return
-
-        # Create a GeneralHookManager to hold MassGen hooks
-        # (We'll convert these to native format)
-        manager = GeneralHookManager()
-
-        # Create mid-stream injection hook with closure-based callback
-        mid_stream_hook = MidStreamInjectionHook()
-
-        # Define the injection callback (A1: unified with the GeneralHookManager
-        # path via the single installer method; native=True only changes log
-        # wording and the non-native debug listing).
-        async def get_injection_content() -> str | None:
-            return await self._build_midstream_injection(agent_id, answers, native=True)
-
-        # Set callback on hook
-        mid_stream_hook.set_callback(get_injection_content)
-
-        # Register mid-stream injection hook
-        manager.register_global_hook(HookType.POST_TOOL_USE, mid_stream_hook)
-
-        # Register high-priority task reminder hook (enabled round-wise when capture is enabled)
-        if self._is_round_learning_capture_enabled():
-            reminder_hook = HighPriorityTaskReminderHook()
-            manager.register_global_hook(HookType.POST_TOOL_USE, reminder_hook)
-
-        # Register media call ledger hook (read_media/generate_media provenance capture)
-        manager.register_global_hook(HookType.POST_TOOL_USE, MediaCallLedgerHook())
-
-        # Register human input hook (shared across all agents)
-        self._ensure_runtime_human_input_hook_initialized()
-        manager.register_global_hook(HookType.POST_TOOL_USE, self._human_input_hook)
-
-        # Register subagent completion hook for background result injection
-        if self._background_subagents_enabled:
-            subagent_hook = SubagentCompleteHook(
-                injection_strategy=self._background_subagent_injection_strategy,
-            )
-
-            # Create a closure that captures agent_id for pending results retrieval
-            def make_pending_getter(aid: str):
-                return lambda: self._get_pending_subagent_results_async(aid)
-
-            subagent_hook.set_pending_results_getter(make_pending_getter(agent_id))
-            manager.register_global_hook(HookType.POST_TOOL_USE, subagent_hook)
-            logger.debug(f"[Orchestrator] Registered SubagentCompleteHook (native) for {agent_id}")
-
-            # Wire background tool delegate so list/status/result/cancel route to subagents
-            if hasattr(agent.backend, "register_background_delegate"):
-                from massgen.subagent.background_delegate import (
-                    SubagentBackgroundDelegate,
-                )
-
-                def _make_call_tool(aid: str):
-                    return lambda tool_name, params: self._call_subagent_mcp_tool_async(
-                        aid,
-                        tool_name,
-                        params,
-                    )
-
-                delegate = SubagentBackgroundDelegate(
-                    call_tool=_make_call_tool(agent_id),
-                    agent_id=agent_id,
-                )
-                agent.backend.register_background_delegate(delegate)
-                logger.debug(f"[Orchestrator] Registered SubagentBackgroundDelegate (native) for {agent_id}")
-
-        # Register background tool completion hook for async tool result injection
-        if hasattr(agent.backend, "get_pending_background_tool_results"):
-            background_tool_hook = BackgroundToolCompleteHook()
-            background_tool_hook.set_completed_jobs_getter(
-                agent.backend.get_pending_background_tool_results,
-            )
-            manager.register_global_hook(HookType.POST_TOOL_USE, background_tool_hook)
-            logger.debug(
-                f"[Orchestrator] Registered BackgroundToolCompleteHook (native) for {agent_id}",
-            )
-        # Register per-round timeout hooks if configured
-        self._register_round_timeout_hooks(agent_id, manager)
-
-        # Register user-configured hooks from agent backend config
-        agent_hooks = agent.backend.config.get("hooks")
-        if agent_hooks:
-            manager.register_hooks_from_config(agent_hooks, agent_id=agent_id)
-
-        # Register PathPermissionManagerHook for PRE_TOOL_USE validation.
-        # Native backends like Copilot need MassGen-level path validation.
-        # Claude Code already handles permissions via add_dirs, so skip it.
-        backend_provider = agent.backend.get_provider_name() if hasattr(agent.backend, "get_provider_name") else ""
-        if backend_provider != "claude_code":
-            _fm = getattr(agent.backend, "filesystem_manager", None)
-            if _fm:
-                _ppm = getattr(_fm, "path_permission_manager", None)
-                if _ppm:
-                    from massgen.filesystem_manager import PathPermissionManagerHook
-
-                    ppm_hook = PathPermissionManagerHook(_ppm)
-                    manager.register_global_hook(HookType.PRE_TOOL_USE, ppm_hook)
-                    logger.debug(
-                        f"[Orchestrator] Registered PathPermissionManagerHook (PRE_TOOL_USE) for {agent_id}",
-                    )
-
-        # Create context factory for hooks
-        def context_factory() -> dict[str, Any]:
-            workspace_path = None
-            filesystem_manager = getattr(agent.backend, "filesystem_manager", None)
-            if filesystem_manager and hasattr(filesystem_manager, "get_current_workspace"):
-                try:
-                    workspace_path = str(filesystem_manager.get_current_workspace())
-                except Exception:
-                    workspace_path = None
-            return {
-                "session_id": getattr(self, "session_id", ""),
-                "orchestrator_id": getattr(self, "orchestrator_id", ""),
-                "agent_id": agent_id,
-                "workspace_path": workspace_path,
-            }
-
-        # Convert to native format using adapter
-        native_config = adapter.build_native_hooks_config(
-            manager,
-            agent_id=agent_id,
-            context_factory=context_factory,
-        )
-
-        # Set native hooks config on backend
-        agent.backend.set_native_hooks_config(native_config)
-        if hasattr(agent.backend, "set_background_wait_interrupt_provider"):
-
-            async def _wait_interrupt_provider(
-                requested_agent_id: str,
-                *,
-                _agent_id: str = agent_id,
-            ) -> dict[str, Any] | None:
-                target_agent_id = requested_agent_id or _agent_id
-                if hasattr(self, "cancellation_manager") and self.cancellation_manager and self.cancellation_manager.is_cancelled:
-                    return {
-                        "interrupt_reason": "turn_cancelled",
-                        "injected_content": None,
-                    }
-
-                runtime_sections = await self._collect_no_hook_runtime_fallback_sections(
-                    target_agent_id,
-                )
-                if not runtime_sections:
-                    return None
-                return {
-                    "interrupt_reason": "runtime_injection_available",
-                    "injected_content": "\n".join(runtime_sections),
-                }
-
-            agent.backend.set_background_wait_interrupt_provider(
-                _wait_interrupt_provider,
-            )
-        logger.info(
-            f"[Orchestrator] Set up native hooks for {agent_id}: " f"PreToolUse={len(native_config.get('PreToolUse', []))}, " f"PostToolUse={len(native_config.get('PostToolUse', []))} hooks",
-        )
+        """Set up native hooks (delegates to MidStreamInjectionHookInstaller)."""
+        self._midstream_injection_hook_installer.setup_native_hooks_for_agent(agent_id, agent, answers)
 
     @classmethod
     def _coerce_answer_content_to_text(cls, content: Any) -> str:
